Match jax, prefer scipy.special.logsumexp
